requirements.txt
Django==4.2
python-magic>=0.4.27
Pillow>=10.0.0
gunicorn>=21.0.0
whitenoise>=6.5.0
psycopg2-binary>=2.9.7